store, jubeatlabaccess: match integer widths to the method encodings
Fix every remaining Objective-C return/parameter width disagreement against the
runtime metadata (rctool objc return-widths now reports 0).

- StoreGenreSelectViewDelegateGenreSelected: the parameter encodes 'Q' (NSUInteger),
  not NSInteger. Narrowed the protocol declaration in StoreGenreSelectView.h and the
  two conformers (StoreMainViewController, StoreMainViewControllerV2), which already
  forward it to -switchToGenre:(NSUInteger).
- jubeatLabAccess currentSize (0x1dae8c): encodes 'Q'; it returns an NSData length, so
  NSUInteger, not NSInteger.
- StoreUtil filePathForMusicID: (0xbb310): the musicID parameter encodes 'I'
  (unsigned int). The binary still formats it with %09d, so keep that specifier with
  an (int) cast to satisfy -Wformat.

+++ Project/StoreUtil.m
@@ -448,9 +448,10 @@ static NSString *const kStorePackProductPrefix = @"jubeat.pack";
 }
 
 /** @ghidraAddress 0xbb310 */
-+ (NSString *)filePathForMusicID:(int)musicID {
-    // "<documents>/%09d.jbt" — the path is returned whether or not the file exists.
-    NSString *name = [[NSString alloc] initWithFormat:@"%09d.jbt", musicID];
++ (NSString *)filePathForMusicID:(unsigned int)musicID {
+    // "<documents>/%09d.jbt" — the binary formats the unsigned id with %09d and returns the path
+    // whether or not the file exists. The cast keeps the %d specifier the binary uses.
+    NSString *name = [[NSString alloc] initWithFormat:@"%09d.jbt", (int)musicID];
     return [JubeatAppDelegate.appDocumentsDirectory stringByAppendingPathComponent:name];
 }
